+++ CHANGELOG.rst
@@ -2,12 +2,14 @@
 CHANGELOG
 =========
 
+
 1.5.0
 ==========
 
 * Added: Support for reading vulns and asset exports into memory
 * Added: Support for graceful back-off
 * Added: Support for tagging API
+* Added: Support for access groups API
 
 1.4.1
 ==========